feat(*): allow whitelisting checks
This commits adds a feature to whitelist specific checks. This is useful
for speeding up builds or excluding useless or broken checks.

Create a "whitelist" file in the config repo with all the checks. Here's
an example:

$ echo "credo" > whitelist
$ echo "gometalinter" >> whitelist

If the "whitelist" file is present, then the check only runs if the
check name is found in the file.