fix(tests): use unique BK meta dir per-test suite
This removes order-dependency in the test suite. Previously (and
incorrectly), metadata persisted between tests. This lead to false
positives for metadata set in each check. This regression covered up the
failure that non-whitelisted checks did not set the title metadata. This
lead to "success" executions inside the builder, but broken on the
application side.