fix(syntax): fix Gemfile issue (#30)
This commit fixes an issue when the code repo contains a Gemfile.
Previously the code required 'bundler/setup' which worked fine in the
tests because there was no Gemfile in the current directory. However the
code doesn't use bundler in the first place. The solution is remove
bundler/setup from check-syntax.